% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_output.R
\name{filter_na}
\alias{filter_na}
\title{Removes lines with NA.}
\usage{
filter_na(simulation, parameter)
}
\arguments{
\item{simulation}{A matrix of real numbers.}

\item{parameter}{A matrix of real numbers with the same number of rows as
\code{simulation}.}
}
\value{
A list containing as first element \code{filtered_simulation} (i.e.
without rows with NA), as second element \code{filtered_parameter} (i.e.
without the rows with the same index as the ones removed from \code{simulation})
and as third element the indexes of the rows of \code{simulation} with NA.
}
\description{
Removes the lines with NA in \code{simulation} and remove the corresponding
rows in \code(parameter).
}
\examples{
simulation <- rnorm(10 * 5)
simulation[sample(1:(10 * 5), 5)] <- NA
simulation <- matrix(simulation, 10, 5)
parameter <- matrix(rnorm(10 * 9), 10, 9)
filter_na(simulation = simulation, parameter = parameter)
}
