% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_output.R
\name{read_clean_output}
\alias{read_clean_output}
\title{Read and clean a simulation file produced by \code{\link[inferenceFitnessLandscape]{simulate_fl}}}
\usage{
read_clean_output(file, keep_param = NULL, two_env = F, ...)
}
\arguments{
\item{file}{Name of the simulation file to import.}

\item{keep_param}{Vector of names corresponding to the names of parameters to keep for the inference.
i.e. the parameters varying in the simulations.}

\item{two_env}{Logical. TRUE if teh sumstat correspond to two environments and FALSE if it
correspond to one. Default = FALSE.}

\item{...}{Options which will be passed to \code{\link[utils]{read.table}}.}
}
\value{
A list. The first element is a matrix of the simulation raw results
(i.e. the fitness of the mutants). Each row correspond to one simulation. The
second element is a matrix of the parameters (only the ones named in \code{keep_param}).
Each row contains the parameters of the corresponding row of "simulation".
}
\description{
Import simulations, removes simulations with NA and keep only given list of parameters.
}
