% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_output.R
\name{remove_wt_fitness}
\alias{remove_wt_fitness}
\title{Removes lines with NA.}
\usage{
remove_wt_fitness(sumstat, two_env = F)
}
\arguments{
\item{sumstat}{A matrix of real numbers. Matrix of summary statistics with each
row containing the raw fitness of a simulation.}

\item{two_env}{Logical. TRUE if teh sumstat correspond to two environments and FALSE if it
correspond to one. Default = FALSE.}
}
\value{
The same matrix as \code{sumstat} minus the first column (and the middle column if two envs) if
all the values are identical.
}
\description{
Removes the first value of each simulations if it is the wild type. These values
are considered as wild type if they are all identical.
}
